refactor: modularize drag and drop
